@@ -18,7 +18,7 @@ classifiers = [
     "Topic :: Software Development :: Libraries",
     "Typing :: Typed",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "httpx[http2]",
     "msgspec",
@@ -32,7 +32,7 @@ dev-dependencies = [
 ]
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 lint.select = [
     "A", "ANN", "ARG", "ASYNC", "B",
     "BLE", "C4", "E", "F", "I",
