@@ -1,5 +1,6 @@
+from collections.abc import AsyncGenerator
 from datetime import date, datetime, timedelta
-from typing import TYPE_CHECKING, AsyncGenerator, Literal, Optional, Tuple
+from typing import TYPE_CHECKING, Literal, Optional
 
 from jkit._base import DATA_OBJECT_CONFIG, DataObject, ResourceObject
 from jkit._network_request import get_json
@@ -38,7 +39,7 @@ class UserEarningRankingData(DataObject, **DATA_OBJECT_CONFIG):
     total_fp_amount_sum: PositiveFloat
     fp_by_creating_amount_sum: PositiveFloat
     fp_by_voting_amount_sum: PositiveFloat
-    records: Tuple[RecordField, ...]
+    records: tuple[RecordField, ...]
 
 
 class UserEarningRanking(ResourceObject):
